backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic[email]==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
